+++ clang-tools-extra/clang-tidy/openmp/OpenMPTidyModule.cpp
@@ -13,6 +13,7 @@
 #include "DoNotModifyLoopVariableCheck.h"
 #include "ExceptionEscapeCheck.h"
 #include "MissingForInParallelDirectiveBeforeForLoopCheck.h"
+#include "SpecifyScheduleCheck.h"
 #include "UnprotectedSharedVariableAccessCheck.h"
 #include "UseDefaultNoneCheck.h"
 
@@ -33,6 +34,8 @@ public:
     CheckFactories
         .registerCheck<MissingForInParallelDirectiveBeforeForLoopCheck>(
             "openmp-missing-for-in-parallel-directive-before-for-loop");
+    CheckFactories.registerCheck<SpecifyScheduleCheck>(
+        "openmp-specify-schedule");
     CheckFactories.registerCheck<UnprotectedSharedVariableAccessCheck>(
         "openmp-unprotected-shared-variable-access");
     CheckFactories.registerCheck<UseDefaultNoneCheck>(